created lda submodule
