+++ mlcvs/lda/__init__.py
@@ -1,0 +1,3 @@
+from .linear_lda import LDA_CV
+from .deep_lda import DeepLDA_CV, ColvarDataset
+from .lda import LDA
